Added delete button for localStorage and added confirm dialogs for deletion actions; got engine and component heat sink totals working; hps and dps reflect laser duration correctly; added rough guess of MWO heat efficiency formula

+++ js/core.js
@@ -127,9 +127,22 @@
 			loadIntoView(mech);
 		};
 
+		self.deleteSaved = function(){
+			var selected = self.selectedSavedMech();
+			if(!selected) return; // can't delete things that aren't real.
+
+			if(confirm('Delete saved loadout "' + self.selectedSavedMech() + '"?')){
+				// Remove key/value from IS
+				localStorage.removeItem(self.selectedSavedMech());
+				mechlab.refreshStorageMechs();
+			}
+		};
+
 		self.clearSaved = function() {
-			localStorage.clear();
-			mechlab.refreshStorageMechs();
+			if(confirm("Delete all saved loadouts?")){
+				localStorage.clear();
+				mechlab.refreshStorageMechs();
+			}
 		};
 
 	}; // end core vm xtor
